% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renv.R
\name{lockfile_pkgs}
\alias{lockfile_pkgs}
\title{If renv is initialized, what packages are in the 'renv.lock'?}
\usage{
lockfile_pkgs()
}
\value{
A character vector of packages recorded in 'renv.lock' or "None".
}
\description{
If renv is initialized, what packages are in the 'renv.lock'?
}
\examples{
lockfile_pkgs()
}
