% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_demo_project}
\alias{template_demo_project}
\title{Returns File Path for Demo Project Template}
\usage{
template_demo_project()
}
\value{
Path to Template YAML file
}
\description{
Returns File Path for Demo Project Template
}
\examples{
cat(readLines(template_demo_project()), sep = "\n")
}
