% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renv.R
\name{use_renv}
\alias{use_renv}
\title{Initialize renv with explicit dependencies}
\usage{
use_renv(path = usethis::proj_get, restart = rlang::is_interactive())
}
\arguments{
\item{path}{Project path}

\item{restart}{Logical: should the R session be restarted after initializing
the project?}
}
\value{
The project path where the renv environment was initialized,
invisibly.
}
\description{
\code{use_renv()} initializes an renv package envrionment for a
project with explicit dependencies. This means that it will only include
packages and their dependencies that are included in a project's
\code{DESCRIPTION} file.
}
